Create fishery from Ecopath with Ecosim (EwE) output data

This function creates fishery observations for an Ecosim scenario.

create_fishery(
  file_path,
  skip_nrows,
  species,
  species_labels,
  ewe_years,
  data_years,
  fleet_num,
  selectivity = NULL,
  CV,
  sample_num,
  waa_path
)

Arguments

file_path

A character string shows path to Ecosim catch output file (e.g., catch_annual.csv) is located.

skip_nrows

interger; the number of lines of the data file to skip before reading data.

species

The species to create fishery for. A vector of column names from the catch_annual.csv.

species_labels

Customized names for species columns.

ewe_years

A vector of years from EwE outputs.

data_years

A vector of years for simulation.

fleet_num

Fleet number.

selectivity

Selectivity at age. A data frame defining selectivity-at-age for a species over time. Columns represent age classes and rows represent years. Provide selectivity data only if fleet_num is greater than 1.

CV

Coefficient of variation of catch. A vector of CVs for individual fleets.

sample_num

Annual sample size. A vector of sample sizes for individual fleets.

waa_path

A character string shows path to Ecosim weight output file (e.g., weight_annual.csv) is located.

Value

A data list that includes catch and composition data from the EwE operating model, and observed catch and composition data for each fleet. Each fleet has time series catch with lognormal error and composition data with sampling error.

Details

The function takes fishery catch data from an Ecosim scenario. The catch data are read in from Ecosim catch_annual.csv and the catch data are in biomass. The function reads in total catch-at-age data of a specific species by age classes and time. EwE already applies a fishery selectivity internally, so no need to reallocate catch if there is only one fleet for the fishery.

See also