Generate frequently used input data for Stock Synthesis 3

Generate frequently used input data for Stock Synthesis 3

generate_ss3(
  file_path,
  r0,
  r0_phase = 1,
  steepness,
  sigmar,
  projection_f,
  projection_catch = NULL,
  sa_data,
  model_year,
  projection_year,
  use_depletion = FALSE,
  depletion_ratio = NULL,
  initial_equilibrium_catch = TRUE,
  settlement_age = 0
)

Arguments

file_path

A file path to a directory where the assessment input and output files will be saved.

r0

Unexploited recruitment from Beverton-Holt stock-recruitment model.

r0_phase

Phase for Unexploited recruitment from Beverton-Holt stock-recruitment model. Default value is 1.

steepness

Steepness from Beverton-Holt stock-recruitment model. Recruitment relative to unfished recruitment at 20% of unfished spawning biomass.

sigmar

Standard deviation of log recruitment.

projection_f

Fishing mortality inputs for projections.

projection_catch

Catch inputs for projections.

sa_data

Simulated stock assessment input data from Rscript/simulation.R.

model_year

A vector of years for model fitting.

projection_year

A vector of years for projection.

use_depletion

A logical value.

depletion_ratio

Ratio of depletion.

initial_equilibrium_catch

TRUE or FALSE.

settlement_age

Settlement age. Default value is age 0.

Value

Outputs from Stock Synthesis 3

Examples

if (FALSE) {
generate_ss3(
  file_path  = here::here("data", "data_rich"),
  r0 = 12, steepness = 0.5, sigmar =0.25,
  projection_f = 0.75, projection_catch = NULL,
  sa_data = sa_data, model_year = 1985:2012,
  projection_year = 2013:2017,
  use_depletion = FALSE, depletion_ratio = NULL
)
}